Bot/requirements.txt
discord.py
python-dotenv
asyncpg
aiohttp
python-dateutil
